+++ src/components/overview.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 import React from "react";
 import ReactDOM from "react-dom";
 import marked from "marked";
@@ -20,8 +21,8 @@ class Overview extends React.Component {
     const playgrounds = Array.prototype.slice.call(this.findPlayground("lang-playground"), 0);
     for (const p in playgrounds) {
       if (playgrounds.hasOwnProperty(p)) {
-        const source = playgrounds[p].textContent;
-        const parent = playgrounds[p].parentNode;
+        const source = playgrounds[p].getElementsByClassName("ecologyCode")[0].textContent;
+        const parent = playgrounds[p].getElementsByClassName("ecologyCode")[0].parentNode;
         ReactDOM.render(
           this.mountContainer(source, parent, true),
           playgrounds[p].parentNode
@@ -32,8 +33,8 @@ class Overview extends React.Component {
       Array.prototype.slice.call(this.findPlayground("lang-playground_norender"), 0);
     for (const p in playgroundsNoRender) {
       if (playgroundsNoRender.hasOwnProperty(p)) {
-        const source = playgroundsNoRender[p].textContent;
-        const parent = playgroundsNoRender[p].parentNode;
+        const source = playgrounds[p].getElementsByClassName("ecologyCode")[0].textContent;
+        const parent = playgrounds[p].getElementsByClassName("ecologyCode")[0].parentNode;
         ReactDOM.render(
           this.mountContainer(source, parent, false),
           playgroundsNoRender[p].parentNode
@@ -41,17 +42,51 @@ class Overview extends React.Component {
       }
     }
   }
+  renderMarkdown() {
+    const { customRenderers, markdown } = this.props;
+    const renderer = new marked.Renderer();
+    const renderers = {
+      code: (code, lang) => {
+        const escape = (html) => {
+          return html
+            .replace(/&/g, "&amp;")
+            .replace(/</g, "&lt;")
+            .replace(/>/g, "&gt;")
+            .replace(/"/g, "&quot;")
+            .replace(/'/g, "&#39;");
+        };
+        // Use regular strings, es6 templates cause spaces to be inserted
+        if (!lang) {
+          return ("<pre><code>" + escape(code) + "</code></pre>");
+        }
+
+        if (lang === "playground" || lang === "playground_norender") {
+          return ("<pre><code class='lang-" + escape(lang) + "'><span class='ecologyCode'>" + escape(code) + "</span></code></pre>");
+        }
+
+        return ("<pre><code class='lang-" + escape(lang) + "'>" + escape(code) + "</code></pre>");
+      },
+      ...customRenderers
+    };
+    Object.assign(renderer, renderers);
+    return marked(markdown, { renderer });
+  }
   render() {
-    const markdown = marked(this.props.markdown);
     return (
-      <div ref="overview" dangerouslySetInnerHTML={{__html: markdown}}>
-      </div>
+      <div ref="overview" dangerouslySetInnerHTML={{__html: this.renderMarkdown()}} />
     );
   }
 }
 
 export default Overview;
 
+Overview.defaultProps = {
+  customRenderers: null
+};
+
 Overview.propTypes = {
   markdown: React.PropTypes.string
+  playgroundtheme: React.PropTypes.string,
+  customRenderers: React.PropTypes.object,
+  scope: React.PropTypes.object
 };