+++ test/client/spec/components/ecology.spec.jsx
@@ -1,12 +1,11 @@
 /**
  * Client tests
  */
-import React from "react/addons";
+import React from "react";
+import TestUtils from "react-addons-test-utils";
 import Ecology from "src/components/ecology";
-
 // Use `TestUtils` to inject into DOM, simulate events, etc.
 // See: https://facebook.github.io/react/docs/test-utils.html
-const TestUtils = React.addons.TestUtils;
 
 describe("components/ecology", function () {