+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Ecology Changelog
 
+## 1.5.2 (2016-07-14)
+
+  * Fix extra padding in codeblocks
+
+## 1.5.1 (2016-07-06)
+
+  * Add classname to `lang-playground` codeblocks
+
+## 1.5.0 (2016-07-04)
+
+  * Add support for custom Marked renderers
+  * Fix development build
+
+## 1.4.0 (2016-06-02)
+
+  * Update component-playground and radium
+
+## 1.3.0 (2016-05-25)
+
+  * Update component-playground
+  * Support React 15
+
 ## 1.2.0 (2016-02-26)
 
   * Ignore source maps in git and npm